refactor testable storage code
