+++ packages/hardhat-router/src/internal/render-template.ts
