+++ packages/hardhat-router/src/tasks/generate-testable.ts
@@ -1,74 +1,49 @@
+import path from 'node:path';
 import { task } from 'hardhat/config';
 import { default as logger } from '@synthetixio/core-utils/utils/io/logger';
-import { default as prompter } from '@synthetixio/core-utils/utils/io/prompter';
-import { readPackageJson } from '@synthetixio/core-utils/utils/misc/npm';
-import {
-  SUBTASK_GENERATE_TESTABLE_STORAGE,
-  SUBTASK_SYNC_SOURCES,
-  TASK_GENERATE_TESTABLE,
-} from '../task-names';
-import { quietCompile } from '../utils/quiet-compile';
+import * as types from '@synthetixio/core-utils/utils/hardhat/argument-types';
+import { parseFullyQualifiedName } from 'hardhat/utils/contract-names';
+import { getSourcesFullyQualifiedNames } from '../internal/contract-helper';
+import { quietCompile } from '../internal/quiet-compile';
+import { SUBTASK_GENERATE_TESTABLE_STORAGE, TASK_GENERATE_TESTABLE } from '../task-names';
 
 export interface DeployTaskParams {
-  noConfirm?: boolean;
-  skipProxy?: boolean;
+  artifacts?: string[];
   debug?: boolean;
   quiet?: boolean;
-  clear?: boolean;
-  alias?: string;
-  modules?: string;
-  instance?: string;
 }
 
 task(TASK_GENERATE_TESTABLE, 'Creates generated test contracts for all storage libraries')
-  .addFlag('noConfirm', 'Skip all confirmation prompts')
+  .addOptionalPositionalParam(
+    'artifacts',
+    'Contract files, names, fully qualified names or folder of contracts to include',
+    ['contracts/storage/'],
+    types.stringArray
+  )
   .addFlag('debug', 'Display debug logs')
   .addFlag('quiet', 'Silence all output')
   .setAction(async (taskArguments: DeployTaskParams, hre) => {
-    const { debug, quiet, noConfirm } = taskArguments;
+    const { artifacts, debug, quiet } = taskArguments;
 
     logger.quiet = !!quiet;
     logger.debugging = !!debug;
-    prompter.noConfirm = !!noConfirm;
-
-    // Do not throw an error on missing package.json
-    // This is so we don't force the user to have the file on tests just for the name
-    try {
-      await logger.title(readPackageJson().name);
-    } catch (err: unknown) {
-      if ((err as { code: string }).code !== 'ENOENT') throw err;
-    }
-
-    await logger.title('DEPLOYER');
 
     await quietCompile(hre, !!quiet);
-    //await hre.run(SUBTASK_LOAD_DEPLOYMENT, { instance: 'general', ...taskArguments });
-    hre.router.deployment = {
-      general: {
-        properties: { completed: false, totalGasUsed: '0' },
-        transactions: {},
-        contracts: {},
-      },
-      sources: {},
-      abis: {},
-    };
 
-    await hre.run(SUBTASK_SYNC_SOURCES, taskArguments);
+    const storageLibs = await getSourcesFullyQualifiedNames(hre, artifacts);
 
-    const storageLibs = Object.values(hre.router.deployment!.general.contracts).filter(
-      ({ isStorageLibrary }) => isStorageLibrary
-    );
-
-    console.log(
-      'storage libs',
-      storageLibs.map((lib) => lib.contractName)
-    );
+    logger.info(`Generating testable storage for ${storageLibs.length} contracts`);
 
     // scan for all storage interfaces
-    for (const storageLibArtifact of storageLibs) {
+    for (const contractFullyQualifiedName of storageLibs) {
+      const { contractName } = parseFullyQualifiedName(contractFullyQualifiedName);
+
       await hre.run(SUBTASK_GENERATE_TESTABLE_STORAGE, {
-        artifact: storageLibArtifact.contractFullyQualifiedName,
-        output: `contracts/modules/test/Testable${storageLibArtifact.contractName}Module.sol`,
+        artifact: contractFullyQualifiedName,
+        output: path.resolve(
+          hre.config.paths.sources,
+          `modules/test/Testable${contractName}Module.sol`
+        ),
       });
     }
   });
